% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BottomFiveZipcodes.R
\name{BottomFiveZipcodes}
\alias{BottomFiveZipcodes}
\title{BottomFiveZipcodes
This function takes a data frame created by the GetDataFromState functiona and shows the zipcodes with the fewest providers}
\usage{
BottomFiveZipcodes(data)
}
\arguments{
\item{the}{data frame created by GetDataFromState}
}
\value{
a graph of the zipcodes with the fewest providers
}
\description{
BottomFiveZipcodes
This function takes a data frame created by the GetDataFromState functiona and shows the zipcodes with the fewest providers
}
\examples{
Data<-GetDataFromState("MA", "Pediatrician")
BottomFiveZipcodes(Data)
}
