% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BottomTaxonomiesZip.R
\name{BottomTaxonomiesZip}
\alias{BottomTaxonomiesZip}
\title{BottomTaxonomiesZip
This function, given a data frame, returns the least frequent provider taxonomies in a zipcode}
\usage{
BottomTaxonomiesZip(data)
}
\arguments{
\item{the}{data frame created with GetDataFromStates}
}
\value{
A visualization of the 5 least frequent provider taxonomies in a zipcode
}
\description{
BottomTaxonomiesZip
This function, given a data frame, returns the least frequent provider taxonomies in a zipcode
}
\examples{
Data<-GetDataFromState("MI", "Primary Care")
BottomTaxonomiesZip(Data)
}
