% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountByZip.R
\name{CountByZip}
\alias{CountByZip}
\title{CountByZip}
\usage{
CountByZip(data)
}
\arguments{
\item{the}{data frame created by GetDataFromState}
}
\value{
the number of providers in a zipcode
}
\description{
This is a function that given the data frame created by GetDataFromState returns a count of the number of providers in a zipcode
}
\examples{
DF1<-GetDataFromState("MA", "family medicine")
CountByZip(DF1)
}
