% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataFromState.R
\name{GetDataFromState}
\alias{GetDataFromState}
\title{GetDataFromState
This function, given a state abbreviation and a provider taxonomy outputs a data fram}
\usage{
GetDataFromState(state, taxonomy)
}
\arguments{
\item{"state}{abbreviation", "taxonomy"}
}
\value{
a data frame
}
\description{
GetDataFromState
This function, given a state abbreviation and a provider taxonomy outputs a data fram
}
\examples{
GetDataFromState("RI", "Primary Care")
}
