% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighProviderNumberZip.R
\name{HighProviderNumberZip}
\alias{HighProviderNumberZip}
\title{HighProviderNumberZip
This outputs a graph of the zip codes with the highest number of providers in a state}
\usage{
HighProviderNumberZip(data)
}
\arguments{
\item{the}{data frame created with GetDataFromState}
}
\value{
a bar graph the the zip codes with the highest number of providers in a taxonomy
}
\description{
HighProviderNumberZip
This outputs a graph of the zip codes with the highest number of providers in a state
}
\examples{
Data<-GetDataFromState("RI", "Mental Health")
HighProviderNumberZip(Data)
}
