% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowProviderNumberZip.R
\name{LowProviderNumberZip}
\alias{LowProviderNumberZip}
\title{LowProviderNumberZip
This function outputs a visualization of the zip codes with the lowest number of providers in a state (with at least one provider minimum in the zip code)}
\usage{
LowProviderNumberZip(data)
}
\arguments{
\item{The}{data frame created by the function GetDataFromState}
}
\value{
a bar graph of the zip codes with the lowest number of providers in a state.
}
\description{
LowProviderNumberZip
This function outputs a visualization of the zip codes with the lowest number of providers in a state (with at least one provider minimum in the zip code)
}
