% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.function.R
\name{NPIcode_taxonomy}
\alias{NPIcode_taxonomy}
\title{NPIcode_taxonomy
This is the base function that created many of our further functions.}
\usage{
NPIcode_taxonomy(zipcode, taxonomy)
}
\description{
NPIcode_taxonomy
This is the base function that created many of our further functions.
}
