% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumProviderInZip.R
\name{NumProviderInZip}
\alias{NumProviderInZip}
\title{NumProvidersInZip
This function takes a zip code and a provider taxonomy and outputs the number of providers in a zip codes}
\usage{
NumProviderInZip(zipcode, taxonomy)
}
\description{
NumProvidersInZip
This function takes a zip code and a provider taxonomy and outputs the number of providers in a zip codes
}
