% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProviderInStateByCounty.R
\name{ProviderInStateByCounty}
\alias{ProviderInStateByCounty}
\title{ProviderInStateByCounty
This fuction creates a data frame of the providers in a state divided by county}
\usage{
ProviderInStateByCounty(state, taxonomy)
}
\arguments{
\item{"state}{abbreviation", "provider taxonomy"}
}
\value{
a data frame
}
\description{
ProviderInStateByCounty
This fuction creates a data frame of the providers in a state divided by county
}
\examples{
ProviderInStateByCounty("CA", "cardiologist")
}
