% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryByCounty.R
\name{SummaryByCounty}
\alias{SummaryByCounty}
\title{SummaryByCounty
This function gives you the number of providers in each county, the number of providers per 1,000 people, and the population of the county}
\usage{
SummaryByCounty(state, taxonomy)
}
\arguments{
\item{a}{data from from provider in the state by county}
}
\value{
a data frame
}
\description{
SummaryByCounty
This function gives you the number of providers in each county, the number of providers per 1,000 people, and the population of the county
}
\examples{
SummaryByCounty("RI", "Cardiologist")
}
