% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopFiveZipcodes.R
\name{TopFiveZipcodes}
\alias{TopFiveZipcodes}
\title{TopFiveZipcodes
This function, given a data frame created with GetDataFromState, creates a vizualtation of the five zipcodes with the most providers}
\usage{
TopFiveZipcodes(data)
}
\arguments{
\item{the}{data frame create by GetDataFromState}
}
\value{
a graph of the five zipcodes with the most providers
}
\description{
TopFiveZipcodes
This function, given a data frame created with GetDataFromState, creates a vizualtation of the five zipcodes with the most providers
}
\examples{
Data<-GetDataFromState("FL", "Addiction")
TopFiveZipcodes(Data)
}
