% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopTaxonomiesZip.R
\name{TopTaxonomiesZip}
\alias{TopTaxonomiesZip}
\title{TopTaxonomiesZip
Given a data set, this function returns the most common provider taxonomies in a zipcode}
\usage{
TopTaxonomiesZip(data)
}
\arguments{
\item{a}{data frame created with GetDataFromState}
}
\value{
a visualization of the most common provider taxonomies
}
\description{
TopTaxonomiesZip
Given a data set, this function returns the most common provider taxonomies in a zipcode
}
\examples{
Data<-GetDataFromState("WI", "Nutrition")
TopTaxonomiesZip(Data)
}
