% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityMapByCounty.R
\name{DensityMapByCounty}
\alias{DensityMapByCounty}
\title{DensityMapByCounty
Maps providers per 1000 people by county}
\usage{
DensityMapByCounty(dat, state)
}
\description{
DensityMapByCounty
Maps providers per 1000 people by county
}
