% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipsFromState.R
\name{ZipsFromState}
\alias{ZipsFromState}
\title{ZipsFromState
Pulls out all the zipcodes from a state}
\usage{
ZipsFromState(state_name)
}
\description{
ZipsFromState
Pulls out all the zipcodes from a state
}
